refactor: update copy and move Intract banner to Bridge carousel
