+++ apps/evm/src/pages/Bridge/components/BannerCarousel/BannerCarousel.style.tsx
@@ -102,6 +102,40 @@ const StyledOnrampBanner = styled(Banner)`
   }}
 `;
 
+const StyledPressableBanner = styled(Banner)`
+  background-image: url(${getImageUrl('cubs-group.svg')});
+  background-repeat: no-repeat;
+  background-size: cover;
+
+  position: relative;
+
+  ${({ theme }) => {
+    return css`
+      background-position: 70% 50%;
+
+      @media ${theme.breakpoints.up('s')} {
+        background-position: 0% 50%;
+      }
+    `;
+  }}
+`;
+
+const StyledIntractBannerImg = styled.img`
+  ${({ theme }) => {
+    return css`
+      height: 4rem;
+
+      @media ${theme.breakpoints.down('s')} {
+        position: absolute;
+        top: 50%;
+        left: 50%;
+        transform: translate(-50%, -50%);
+        opacity: 0.2;
+      }
+    `;
+  }}
+`;
+
 const StyledOnrampGraphic = styled(OnrampGraphic)`
   ${({ theme }) => {
     return css`
@@ -130,5 +164,7 @@ export {
   StyledOnrampGraphic,
   StyledBannerContent,
   StyledBanner,
-  StyledBannerImg
+  StyledBannerImg,
+  StyledPressableBanner,
+  StyledIntractBannerImg
 };
