+++ apps/evm/src/pages/Bridge/components/BannerCarousel/BannerCarousel.tsx
@@ -7,6 +7,7 @@ import { useTheme } from 'styled-components';
 import { EcosystemBanner } from './EcosystemBanner';
 import { StyledCarouselWrapper, StyledSlider } from './BannerCarousel.style';
 import { OnrampBanner } from './OnrampBanner';
+import { IntractBanner } from './IntractBanner';
 
 const settings: Settings = {
   dots: true,
@@ -44,6 +45,7 @@ const BannerCarousel = ({ onPressOnrampBanner, onPressEcosystemBanner }: BannerC
       <StyledSlider {...settings} arrows={isDesktop}>
         <OnrampBanner onPress={onPressOnrampBanner} />
         <EcosystemBanner onPress={onPressEcosystemBanner} />
+        <IntractBanner />
       </StyledSlider>
     </StyledCarouselWrapper>
   );
