+++ apps/evm/src/pages/Fusion/Fusion.tsx
@@ -6,15 +6,7 @@ import { useLocalStorage } from '@uidotdev/usehooks';
 import { Geoblock, Main } from '../../components';
 import { LocalStorageKey } from '../../constants';
 
-import {
-  AllUsersLeaderboard,
-  Challenges,
-  Dashboard,
-  Info,
-  IntractBanner,
-  PartnersSection,
-  QuestUsersLeaderboard
-} from './components';
+import { AllUsersLeaderboard, Challenges, Dashboard, Info, PartnersSection, QuestUsersLeaderboard } from './components';
 import { StyledUpdateMark } from './Fusion.style';
 
 const Fusion = () => {
@@ -60,7 +52,6 @@ const Fusion = () => {
           <TabsItem key='dashboard' title='Dashboard'>
             <Flex direction='column' gap='2xl'>
               <Dashboard />
-              <IntractBanner />
               <Challenges />
               <PartnersSection />
             </Flex>
