--- apps/evm/src/pages/Fusion/components/IntractBanner/Banner.tsx
@@ -1,9 +1,0 @@
-import { Card, CardProps } from '@gobob/ui';
-
-type BannerProps = CardProps;
-
-const Banner = (props: BannerProps) => {
-  return <Card direction='row' paddingX={{ base: '2xl', s: '6xl' }} paddingY='5xl' {...props} />;
-};
-
-export { Banner };
