--- apps/evm/src/pages/Fusion/components/IntractBanner/index.tsx
@@ -1,1 +1,0 @@
-export { IntractBanner } from './IntractBanner';
