+++ apps/evm/src/pages/Fusion/components/index.tsx
@@ -2,7 +2,6 @@ export * from './Challenges';
 export * from './Dashboard';
 export * from './Info';
 export * from './Leaderboard';
-export * from './IntractBanner';
 export * from './LoginSignUp';
 export * from './PartnersSection';
 export * from './ProjectStatus';
